Estacionamiento/python/requirements.txt
requests>=2.28.0
mysql-connector-python>=8.0.33
pandas>=1.5.0
matplotlib>=3.6.0
